requirements-dev.txt
pytest==7.0.0
setuptools==67.7.2
wheel==0.40.0
black==22.3.0
flake8==4.0.1
isort==5.12.0

streamlit==1.28.2
requests==2.31.0
jinja2==3.1.2
boto3==1.33
Pillow==10.1.0
numpy==1.26.2
click==8.1.7
PyYAML==6.0
